+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.1
+* **BREAKING**: Changed `icons` parameter from `List<IconData>` to `List<Widget>` to support any widget as an icon.
+* **BREAKING**: Changed `icon` parameter in `NavbarItemWidget` from `IconData` to `Widget`.
+* Added `selectedColor` and `unselectedColor` parameters to `NavbarWidget` for better customization.
+
 ## 2.0.0
 * Renamed package to "LiquidGlass NavBar"
 * Updated horizontal padding parameter to be optional